requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
facenet-pytorch @ file:///C:/Users/Arunabha/Desktop/Projects/deepfake-backend/input/facenet_pytorch-2.2.7-py3-none-any.whl#sha256=dc48e19aaf28e9f9e178424a777b3d2caeab415872e0dbf93ac586da9084539d
fastapi==0.115.8
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
